Added mobile views, styling corrections
