--- app/assets/stylesheets/Landing.css
@@ -1,26 +1,0 @@
-#landing_page .gradient {
-  background: -moz-linear-gradient(45deg, rgba(42, 27, 161, 0.7), rgba(29, 210, 177, 0.7) 100%);
-  background: -webkit-linear-gradient(45deg, rgba(42, 27, 161, 0.7), rgba(29, 210, 177, 0.7) 100%);
-  background: -webkit-gradient(linear, 45deg, from(rgba(42, 27, 161, 0.7)), to(rgba(29, 210, 177, 0.7)));
-  background: -o-linear-gradient(45deg, rgba(42, 27, 161, 0.7), rgba(29, 210, 177, 0.7) 100%);
-  background: linear-gradient(to 45deg, rgba(42, 27, 161, 0.7), rgba(29, 210, 177, 0.7) 100%);
-}
-
-#landing_page .view {
-  background-image: url('aerial-view-architecture-buildings-248867.jpg');
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-position: center center;
-  height: 100vh;
-}
-
-#landing_page h6 {
-  line-height: 1.7;
-}
-
-
-@media (max-width: 768px) {
-  #landing_page .navbar:not(.top-nav-collapse) {
-      background: #424f95 !important;
-  }
-}
