+++ app/assets/stylesheets/Landing.scss
@@ -1,0 +1,48 @@
+.z-depth-1{
+  /* padding-top: 10vh; */
+  /* margin-top: 5vh; */
+  /* border: 5px;
+  border-style: solid;
+  border-color: red; */
+}
+.intro-content{
+  /* padding-top: 15vh; */
+  /* margin-top: 10vh; */
+  /* border: 5px;
+  border-style: solid;
+  border-color: red; */
+}
+
+/* .mt-xl-5{
+  border: 5px;
+  border-style: solid;
+  border-color: blue; */
+
+
+#landing_page .gradient {
+    background: linear-gradient(45deg, rgba(42, 27, 161, 0.7), rgba(29, 210, 177, 0.7) 100%);
+  }
+ 
+  #landing_page .view {
+    background-image: url('aerial-view-architecture-buildings-248867.jpg');
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center center;
+    height: 100vh;
+  }
+
+  #landing_page h6 {
+    line-height: 1.7;
+  }
+
+  .z-depth-2{
+    border-style: solid;
+    border-color: red;
+    border: 2px;
+  }
+
+  @media (max-width: 768px) {
+    #landing_page .navbar:not(.top-nav-collapse) {
+        background: #424f95 !important;
+    }
+  }
