+++ app/assets/stylesheets/Locations.css
@@ -4,14 +4,6 @@
 }
 
 
-.SideNavHeader{
-  height: 20vh;
-  background-image: url("./break_in_logos-01.png");
-  background-size: 12vw;
-  background-repeat: no-repeat;
-  background-position: center;
-}
-
 .bottomModule{
   width: 15vw;
 }
