+++ app/assets/stylesheets/NavBar.css
@@ -8,6 +8,14 @@
     background: #424f95 !important;
   }
 
+  .SideNavHeader{
+    height: 20vh;
+    background-image: url("./break_in_logos-01.png");
+    background-size: 10em;
+    background-repeat: no-repeat;
+    background-position: center;
+  }
+
   @media (max-width: 768px) {
     #Nav .navbar:not(.top-nav-collapse) {
         background: #424f95 !important;
