+++ app/javascript/bundles/App/Components/Landing.js
@@ -1,25 +1,7 @@
 import React from 'react';
-import { Mask, Row, Col, Button, View, Container} from 'mdbreact';
 import  {Carousel, CarouselCaption, CarouselInner, CarouselItem} from 'mdbreact';
+import { Mask, Row, Col, Button, View, Container} from 'mdbreact';
 
-
-//DEFINES WHAT HAPPEN WHEN USER CLICKS LOG IN
-const handleLogIn = () => {
-  let link = document.createElement('a');
-  link.setAttribute('href', '/users/sign_in');
-  link.setAttribute('rel', 'nofollow');
-  document.body.appendChild(link);
-  link.click();
-}
-
-//DEFINES WHAT HAPPENS WHEN USER CLICKS SIGN UP
-const handleCreateAccount = () => {
-  let link = document.createElement('a');
-  link.setAttribute('href', '/users/sign_up');
-  link.setAttribute('rel', 'nofollow');
-  document.body.appendChild(link);
-  link.click();
-}
 
 class LandingPage extends React.Component {
 
@@ -41,11 +23,11 @@ class LandingPage extends React.Component {
                 <div className="white-text text-center text-md-left col-md-6 mt-xl-5 mb-5">
                   <h1 className="h1-responsive font-weight-bold mt-sm-5">Welcome to Break-In App </h1>
                   <hr className="hr-light"/>
-                  <h6 className="my-5">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga nesciunt
+                  <h6 className="mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga nesciunt
                   dolorum nulla magnam veniam sapiente, fugiat! Commodi sequi non animi ea dolor molestiae
                   iste.</h6>
-                <Button onClick={ handleCreateAccount } color="white">Get Started</Button>
-                <Button onClick={ handleLogIn } outline color="white">Have an account? Log in</Button>
+                  <Button color="white">Download</Button>
+                  <Button outline color="white">Learn More</Button>
                 </div>
                 <Col md="6" xl="5" className="mt-xl-5">
                   <img src="https://mdbootstrap.com/img/Mockups/Transparent/Small/admin-new.png" className="img-fluid" alt="landingImg"/>
@@ -60,8 +42,8 @@ class LandingPage extends React.Component {
                   <h6 className="mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga nesciunt
                   dolorum nulla magnam veniam sapiente, fugiat! Commodi sequi non animi ea dolor molestiae
                   iste.</h6>
-                  <Button onClick={ handleCreateAccount } color="white">Get Started</Button>
-                  <Button onClick={ handleLogIn } outline color="white">Have an account? Log in</Button>
+                  <Button color="white">Download</Button>
+                  <Button outline color="white">Learn More</Button>
                 </div>
                 <Col md="6" xl="5" className="mt-xl-5">
                   <img src="https://mdbootstrap.com/img/Mockups/Transparent/Small/admin-new.png" className="img-fluid" alt="landingImg"/>
@@ -76,8 +58,8 @@ class LandingPage extends React.Component {
                   <h6 className="mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga nesciunt
                   dolorum nulla magnam veniam sapiente, fugiat! Commodi sequi non animi ea dolor molestiae
                   iste.</h6>
-                  <Button onClick={ handleCreateAccount } color="white">Get Started</Button>
-                  <Button onClick={ handleLogIn } outline color="white">Have an account? Log in</Button>
+                  <Button color="white">Download</Button>
+                  <Button outline color="white">Learn More</Button>
                 </div>
                 <Col md="6" xl="5" className="mt-xl-5">
                   <img src="https://mdbootstrap.com/img/Mockups/Transparent/Small/admin-new.png" className="img-fluid" alt="landingImg"/>
@@ -92,8 +74,8 @@ class LandingPage extends React.Component {
                   <h6 className="mb-4">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga nesciunt
                   dolorum nulla magnam veniam sapiente, fugiat! Commodi sequi non animi ea dolor molestiae
                   iste.</h6>
-                  <Button onClick={ handleCreateAccount } color="white">Get Started</Button>
-                  <Button onClick={ handleLogIn } outline color="white">Have an account? Log in</Button>
+                  <Button color="white">Download</Button>
+                  <Button outline color="white">Learn More</Button>
                 </div>
                 <Col md="6" xl="5" className="mt-xl-5">
                   <img src="https://mdbootstrap.com/img/Mockups/Transparent/Small/admin-new.png" className="img-fluid" alt="landingImg"/>
