+++ app/javascript/bundles/App/Components/Maps/Places.js
@@ -45,40 +45,39 @@ class Places extends Component {
 		const bottomList = (
 		<div className="bottomList">
 		<div className={classes.bottomList}>
-		<div className="reportHeader">
-			<div className="reportHeaderText">Send a report</div>
-		</div>
+			<div className="reportHeader">
+				<div className="reportHeaderText">Send a report</div>
+			</div>
 			<div className="row">
-
-    <div className="col">
-		<Button className="reportButton1" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Theft</Button>
-    </div>
-    <div className="col">
-		<Button className="reportButton2" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Arson</Button>
-    </div>
-		<div className="col">
-		<Button className="reportButton3" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Hit & Run</Button>
-    </div>
-		<div className="col">
-		<Button className="reportButton4" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Assault</Button>
-    </div>
-		<div className="w-100"></div>
-		<div className="col">
-		<Button className="reportButton5" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Robbery</Button>
-    </div>
-		<div className="col">
-		<Button className="reportButton6" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Drug Deal</Button>
-    </div>
-    <div className="col">
-		<Button className="reportButton7" fullWidth='false' onClick={this.toggleDrawer('bottom', true)}>Murder</Button>
-    </div>
-    <div className="col">
-		<Button className="reportButton8" fullWidth='false' onClick={this.toggleDrawer('bottom', true)}>Camera</Button>
-    </div>
+				<div className="col">
+				<Button className="reportButton1" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Theft</Button>
+				</div>
+    		<div className="col">
+				<Button className="reportButton2" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Arson</Button>
+				</div>
+				<div className="col">
+				<Button className="reportButton3" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Hit & Run</Button>
+				</div>
+				<div className="col">
+				<Button className="reportButton4" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Assault</Button>
+				</div>
+				<div className="w-100"></div>
+				<div className="col">
+				<Button className="reportButton5" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Robbery</Button>
+				</div>
+				<div className="col">
+				<Button className="reportButton6" fullWidth='true' onClick={this.toggleDrawer('bottom', true)}>Drug Deal</Button>
+				</div>
+				<div className="col">
+				<Button className="reportButton7" fullWidth='false' onClick={this.toggleDrawer('bottom', true)}>Murder</Button>
+				</div>
+				<div className="col">
+				<Button className="reportButton8" fullWidth='false' onClick={this.toggleDrawer('bottom', true)}>Camera</Button>
+				</div>
+			</div>
 		</div>
-		</div>
-		</div>
-		);
+	</div>
+);
