+++ app/views/layouts/application.html.erb
@@ -2,6 +2,8 @@
 <html>
   <head>
     <title>PitchDay</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
     <%= javascript_pack_tag 'App' %>
